+++ app/src/main/java/org/sopt/at/presentation/main/navigation/MainNavigator.kt
@@ -7,7 +7,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
-import org.sopt.at.navigation.MainBottomTab
+import org.sopt.at.presentation.main.MainBottomTab
 import org.sopt.at.presentation.home.navigation.Home
 import org.sopt.at.presentation.home.navigation.navigateToHome
 import org.sopt.at.presentation.live.navigation.navigateToLive
